feat: replace file-edit with enhanced multi-edit implementation
- Supports both old API (old_text/new_text) and new API (edits array) for backward compatibility
- Adds multiple sequential edits with occurrence validation
- Includes dry run mode for previewing changes
- Provides comprehensive diff context for UI display
- Maintains atomic operations (all edits succeed or none do)
- Comprehensive test coverage for all functionality
